Get sub-region vaccination data

Runs a GET request of sub-region vaccination data from the COVID-19 tracker API, and returns parsed data. The dates argument specifies the time frame of the data: "current" (the default; latest report for each sub-region), "recent" (15 most recent reports for each sub-region), and "all" (returns all reports for one or more sub-regions specified by the subregion_code argument). To get a list of available sub-regions, use the function get_subregions().

get_subregion_vaccination_data(
  dates = c("current", "recent", "all"),
  subregion_code = NULL
)

Arguments

dates

One of "current", "recent", or "all" to specify the time frame of the reports returned. If choosing "all" reports, must also provide one or more sub-region codes.

subregion_code

One or more sub-region codes. Returns all reports for those sub-regions (even if dates is not "all")

Value

A data frame with one row per sub-region report.

Details

Note that sub-region vaccination data is only for select provinces and territories. Also the percentages reported differ between percent of total population, and percent of eligible population. See the API documentation for more details: https://api.covid19tracker.ca/docs/1.0/vaccinations.

Examples

get_subregion_vaccination_data()
#> # A tibble: 806 × 12
#>    last_updated        code  date       total_dose_1 percent_dose_1
#>    <dttm>              <chr> <date>            <int>          <dbl>
#>  1 2021-12-31 00:00:00 ON001 2021-12-10        94206          0.802
#>  2 2021-12-31 00:00:00 ON002 2021-12-10        16667          0.742
#>  3 2021-12-31 00:00:00 ON003 2021-12-10        42626          0.763
#>  4 2021-12-31 00:00:00 ON004 2021-12-10        32777          0.782
#>  5 2021-12-31 00:00:00 ON005 2021-12-10        33222          0.795
#>  6 2021-12-31 00:00:00 ON006 2021-12-10        38548          0.807
#>  7 2021-12-31 00:00:00 ON007 2021-12-10        26492          0.732
#>  8 2021-12-31 00:00:00 ON008 2021-12-10        94323          0.762
#>  9 2021-12-31 00:00:00 ON009 2021-12-10        61440          0.757
#> 10 2021-12-31 00:00:00 ON010 2021-12-10        43295          0.777
#> # … with 796 more rows, and 7 more variables: source_dose_1 <chr>,
#> #   total_dose_2 <int>, percent_dose_2 <dbl>, source_dose_2 <chr>,
#> #   total_dose_3 <int>, percent_dose_3 <dbl>, source_dose_3 <chr>
get_subregion_vaccination_data("recent")
#> # A tibble: 1,470 × 9
#>    last_updated        date       code  total_dose_1 percent_dose_1
#>    <dttm>              <date>     <chr>        <int>          <dbl>
#>  1 2022-01-03 00:00:00 2021-12-20 AB001         4580          0.729
#>  2 2022-01-03 00:00:00 2021-12-20 AB002         6322          0.758
#>  3 2022-01-03 00:00:00 2021-12-20 AB003         3933          0.582
#>  4 2022-01-03 00:00:00 2021-12-20 AB004        11845          0.714
#>  5 2022-01-03 00:00:00 2021-12-20 AB005        15153          0.587
#>  6 2022-01-03 00:00:00 2021-12-20 AB006         9526          0.501
#>  7 2022-01-03 00:00:00 2021-12-20 AB007         6859          0.618
#>  8 2022-01-03 00:00:00 2021-12-20 AB008         2721          0.425
#>  9 2022-01-03 00:00:00 2021-12-20 AB009        19026          0.686
#> 10 2022-01-03 00:00:00 2021-12-20 AB010         2238          0.642
#> # … with 1,460 more rows, and 4 more variables: source_dose_1 <chr>,
#> #   total_dose_2 <int>, percent_dose_2 <dbl>, source_dose_2 <chr>
get_subregion_vaccination_data("all")
#> Error in get_subregion_vaccination_data("all"): Must specify sub-region(s) to return all vaccination reports.