Get vaccination data

Runs a GET request of vaccination data from the COVID-19 tracker API, and returns parsed data. Data may be returned as type = "summary" (the most recent data) or type = "reports" (day-to-day reports). Via the split argument, data may be "overall" (all provinces/territories combined), by "province", or by "region". Alternatively, provide one or more two-letter codes (e.g. "AB") to province to return reports for specific provinces, or one or more numeric region codes (e.g. "1204") to return specific health regions.

get_vaccination_data(
  type = c("summary", "reports"),
  split = c("overall", "province", "region"),
  province = NULL,
  region = NULL,
  fill_dates = NULL,
  stat = NULL,
  date = NULL,
  after = NULL,
  before = NULL
)

Arguments

type

One of "summary" (most recent data) or "reports" (day-to-day data).

split

One of "overall", "province", or "region" to specify how the data is split. An "overall" summary or report gives cumulative numbers across Canada. Splitting by "province" returns data for all provinces/territories. Splitting by "region" is only available for "summary" data, and returns data for all health regions.

province

One or more province/territory codes ("AB", "BC", "MB", "NB", "NL", "NS", "NT", "NU", "ON", "PE", "QC", "SK", "YT") to get reports. Upper, lower and mixed case strings are accepted.

region

One or more health region IDs to get reports. Numeric and character values are accepted.

fill_dates

When TRUE, the response fills in any missing dates with blank entries.

stat

Returns only the specified statistics, e.g. "cases".

date

Returns reports from only the specified date.

after

Returns reports from only on or after the specified date.

before

Returns reports from only on or before the specified date.

Value

A data frame containing the vaccinations data. Includes a province variable if data is split by province, and a hr_uid variable if data is split by health region.

Examples

get_vaccination_data()
#> # A tibble: 1 × 10
#>   last_updated        latest_date change_vaccinations change_vaccinated
#>   <dttm>              <date>                    <int>             <int>
#> 1 2022-01-03 00:00:00 2022-01-03                 8715               789
#> # … with 6 more variables: change_boosters_1 <int>,
#> #   change_vaccines_distributed <int>, total_vaccinations <int>,
#> #   total_vaccinated <int>, total_boosters_1 <int>,
#> #   total_vaccines_distributed <int>
get_vaccination_data(split = "province")
#> # A tibble: 13 × 11
#>    province last_updated        date       change_vaccinations change_vaccinated
#>    <chr>    <dttm>              <date>                   <int>             <int>
#>  1 ON       2022-01-03 00:00:00 2022-01-03                   0                 0
#>  2 QC       2022-01-03 00:00:00 2022-01-03                   0                 0
#>  3 NS       2022-01-03 00:00:00 2022-01-03                   0                 0
#>  4 NB       2022-01-03 00:00:00 2022-01-03                7460               257
#>  5 MB       2022-01-03 00:00:00 2022-01-03                   0                 0
#>  6 BC       2022-01-03 00:00:00 2022-01-03                   0                 0
#>  7 PE       2022-01-03 00:00:00 2022-01-03                   0                 0
#>  8 SK       2022-01-03 00:00:00 2022-01-03                   0                 0
#>  9 AB       2022-01-03 00:00:00 2022-01-03                   0                 0
#> 10 NL       2022-01-03 00:00:00 2022-01-03                1255               532
#> 11 NT       2022-01-03 00:00:00 2022-01-03                   0                 0
#> 12 YT       2022-01-03 00:00:00 2022-01-03                   0                 0
#> 13 NU       2022-01-03 00:00:00 2022-01-03                   0                 0
#> # … with 6 more variables: change_boosters_1 <int>,
#> #   change_vaccines_distributed <int>, total_vaccinations <int>,
#> #   total_vaccinated <int>, total_boosters_1 <int>,
#> #   total_vaccines_distributed <int>
get_vaccination_data()
#> # A tibble: 1 × 10
#>   last_updated        latest_date change_vaccinations change_vaccinated
#>   <dttm>              <date>                    <int>             <int>
#> 1 2022-01-03 00:00:00 2022-01-03                 8715               789
#> # … with 6 more variables: change_boosters_1 <int>,
#> #   change_vaccines_distributed <int>, total_vaccinations <int>,
#> #   total_vaccinated <int>, total_boosters_1 <int>,
#> #   total_vaccines_distributed <int>
get_vaccination_data()
#> # A tibble: 1 × 10
#>   last_updated        latest_date change_vaccinations change_vaccinated
#>   <dttm>              <date>                    <int>             <int>
#> 1 2022-01-03 00:00:00 2022-01-03                 8715               789
#> # … with 6 more variables: change_boosters_1 <int>,
#> #   change_vaccines_distributed <int>, total_vaccinations <int>,
#> #   total_vaccinated <int>, total_boosters_1 <int>,
#> #   total_vaccines_distributed <int>
get_vaccination_data()
#> # A tibble: 1 × 10
#>   last_updated        latest_date change_vaccinations change_vaccinated
#>   <dttm>              <date>                    <int>             <int>
#> 1 2022-01-03 00:00:00 2022-01-03                 8715               789
#> # … with 6 more variables: change_boosters_1 <int>,
#> #   change_vaccines_distributed <int>, total_vaccinations <int>,
#> #   total_vaccinated <int>, total_boosters_1 <int>,
#> #   total_vaccines_distributed <int>